:recycle: refactor: ajustando services e criando validações

+++ Squad33/src/main/java/com/squad33/api/sevice/impl/CursoServiceImpl.java
@@ -5,7 +5,9 @@ import java.util.Optional;
 
 import com.squad33.api.sevice.ICursoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.squad33.api.models.Curso;
 import com.squad33.api.repositories.CursoRepository;
@@ -19,8 +21,20 @@ public class CursoServiceImpl implements ICursoService {
 	public Curso save(Curso curso) {
 		return repository.save(curso);
 	}
+	
+	public Curso update(Integer id) {
+		if(!repository.findById(id).isPresent()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+		
+		Curso curso = repository.findById(id).orElse(null);
+		return save(curso);
+	}
 
 	public void deleteById(Integer id) {
+		if(!repository.findById(id).isPresent()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
 		repository.deleteById(id);
 	}
 	
@@ -34,7 +48,7 @@ public class CursoServiceImpl implements ICursoService {
 
 	@Override
 	public Curso findById(Integer id) {
-		return repository.findById(id).orElse(null);
+		return repository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
 	}
 
 	public Optional<Curso> FindById(Integer id) {
